# backend/api/docs/swagger.yaml
basePath: /
definitions:
...
  echo-backend_internal_models_user.BudgetMonitorConfig:
    properties:
      enabled:
        type: boolean
      enforceCostCap:
        type: boolean
      enforceMaxSteps:
        type: boolean
      enforceTimeout:
        type: boolean
      maxCostUsd:
        type: number
      maxDurationMs:
        type: integer
      maxSteps:
        type: integer
    type: object
  echo-backend_internal_models_user.ContextOptimizationConfig:
    properties:
      compactionThresholdRatio:
        type: number
      enableAutoCompaction:
        type: boolean
      enablePrefixCachingLayout:
        type: boolean
      enabled:
        type: boolean
      keepLastTurnsCount:
        type: integer
    type: object
  echo-backend_internal_models_user.HarnessFeatureToggles:
    properties:
      budgetMonitor:
        $ref: '#/definitions/echo-backend_internal_models_user.BudgetMonitorConfig'
      contextOptimization:
        $ref: '#/definitions/echo-backend_internal_models_user.ContextOptimizationConfig'
      hitlGuard:
        $ref: '#/definitions/echo-backend_internal_models_user.HitlGuardConfig'
      loopDetection:
        $ref: '#/definitions/echo-backend_internal_models_user.LoopDetectionConfig'
      systemNotices:
        $ref: '#/definitions/echo-backend_internal_models_user.SystemNoticesConfig'
    type: object
  echo-backend_internal_models_user.HitlGuardConfig:
    properties:
      enabled:
        type: boolean
      protectedTools:
        items:
          type: string
        type: array
      ttlMinutes:
        type: integer
    type: object
  echo-backend_internal_models_user.LoopDetectionConfig:
    properties:
      enableCosineSimilarity:
        type: boolean
      enableExactMatch:
        type: boolean
      enabled:
        type: boolean
      maxConsecutiveIdenticalCalls:
        type: integer
      similarityThreshold:
        type: number
      windowSize:
        type: integer
    type: object
  echo-backend_internal_models_user.SystemNoticesConfig:
    properties:
      emitBudgetWarnings:
        type: boolean
      emitCompactionNotices:
        type: boolean
      emitLoopWarnings:
        type: boolean
      emitPacingWarnings:
        type: boolean
      enabled:
        type: boolean
    type: object
  echo-backend_internal_models_user.UserPreferences:
    properties:
      api_key:
        example: ""
        type: string
      base_url:
        example: https://opencode.ai/zen/go/v1
        type: string
      default_features:
        example:
        - web-browsing
        - code-interpreter
        items:
          type: string
        type: array
      default_mode:
        example: agent
        type: string
      default_model:
        example: gpt-4o
        type: string
      default_skills:
        example:
        - python
        - research
        items:
          type: string
        type: array
      harness_toggles:
        $ref: '#/definitions/echo-backend_internal_models_user.HarnessFeatureToggles'
      has_api_key:
        type: boolean
      provider_type:
        example: opencode-go
        type: string
      user_id:
        example: 1
        type: integer
    type: object
...
  internal_handler_settings.UpdateSettingsRequest:
    properties:
      api_key:
        type: string
      base_url:
        type: string
      default_features:
        items:
          type: string
        type: array
      default_mode:
        type: string
      default_model:
        type: string
      default_skills:
        items:
          type: string
        type: array
      harness_toggles:
        $ref: '#/definitions/echo-backend_internal_models_user.HarnessFeatureToggles'
      keep_api_key:
        type: boolean
      provider_type:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@echo.app
    name: Echo API Support
    url: https://echo.app/support
  description: Core API Server for Echo Platform providing Auth, Chat, Sessions, Models,
    Settings, Admin, Memory, and System Services.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://echo.app/terms
  title: Echo Backend API
  version: "1.0"
paths:
...
  /api/v1/settings:
    get:
      description: Returns the authenticated user's preferences
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/echo-backend_internal_models_user.UserPreferences'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user settings
      tags:
      - Settings
    put:
      consumes:
      - application/json
      description: Updates the authenticated user's preferences
      parameters:
      - description: Settings payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler_settings.UpdateSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/echo-backend_internal_models_user.UserPreferences'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update user settings
      tags:
      - Settings
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authorization header (Format: Bearer <JWT>).'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"